\name{CarrollPack-package}
\alias{CarrollPack-package}
\alias{CarrollPack}
\docType{package}
\title{
Data Mines a User-Supplied Dataset
}
\description{
Function runs OLS regressions for every possible combination of independnet variables given by the user, and returns a matrix containing the coefficient estimates and R^2 values from each of these linear models. 
}
\details{
\tabular{ll}{
Package: \tab CarrollPack\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-11\cr
License: \tab What license is it under?\cr
Depends: \tab methods\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Thomas Carroll

Maintainer: Thomas Carroll <thomasscarroll89@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
